design: placeInfo component 분리 (#4)
